+++ controllers/contactController.js
@@ -1,0 +1,84 @@
+const express = require("express");
+
+const contactsService = require("../models/contacts");
+const { HttpError } = require("../helpers");
+const { contactAddSchema } = require("../schemas");
+
+const router = express.Router();
+
+const listContacts = async (req, res, next) => {
+  try {
+    const result = await contactsService.listContacts();
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const getContactById = async (req, res, next) => {
+  try {
+    console.log(req.params);
+    const { contactId } = req.params;
+    const result = await contactsService.getContactById(contactId);
+    if (!result) {
+      throw HttpError(404, `Contact not found with id: ${contactId}`);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const removeContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contactsService.removeContact(contactId);
+    if (!result) {
+      throw HttpError(404, `Contact not found with id: ${contactId}`);
+    }
+    res.json({ message: "Contact deleted successfully" });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const addContact = async (req, res, next) => {
+  try {
+    const { error } = contactAddSchema.validate(req.body);
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+    const result = await contactsService.addContact(req.body);
+    res.status(201).json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const updateContact = async (req, res, next) => {
+  try {
+    const { error } = contactAddSchema.validate(req.body);
+
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+    const { contactId } = req.params;
+    const result = await contactsService.updateContact(contactId, req.body);
+
+    if (!result) {
+      throw HttpError(404, `Contact with id: ${contactId} not found`);
+    }
+
+    res.status(200).json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+router.get("/", listContacts);
+router.get("/:contactId", getContactById);
+router.delete("/:contactId", removeContact);
+router.post("/", addContact);
+router.put("/:contactId", updateContact);
+
+module.exports = router;
