+++ models/contacts.json
@@ -1,9 +1,9 @@
 [
   {
-    "email": "nulla.ante@vestibul.co.uk",
     "id": "AeHIrLTr6JkxGE6SN-0Rw",
-    "name": "Allen Raymond",
-    "phone": "(992) 914-3792"
+    "name": "Thomas Lucas",
+    "email": "nec@Nulla.com",
+    "phone": "(704) 398-79934455"
   },
   {
     "email": "dui.in@egetlacus.ca",
@@ -64,5 +64,17 @@
     "name": "Thomas Lucas",
     "email": "nec@Nulla.com",
     "phone": "(704) 398-79934455"
+  },
+  {
+    "id": "i3HzcY4GlHUPPuoTtkIx3",
+    "name": "Igor Bere",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685"
+  },
+  {
+    "id": "r175Rlu57Ugn2CH3RIK2T",
+    "name": "Igor Bere",
+    "email": "dui.in@egetlacus.ca",
+    "phone": "(294) 840-6685"
   }
 ]
