+++ schemas/index.js
@@ -1,0 +1,5 @@
+const contactAddSchema = require("./contactAdd");
+
+module.exports = {
+  contactAddSchema,
+};
